requirements.txt
PyYAML==6.0.1
termcolor==2.4.0
luno-python==0.0.10
redis==5.0.1
python-dotenv==1.0.1
scipy==1.11.3
websockets==12.0